+++ CHANGELOG.md
@@ -10,4 +10,4 @@
 
 ## 1.0.2
 * Add initIndex,initOffset and initOffsetBasedOnBottom
-* Add Element Position Callback
+* Add Element Positions Callback